task 1 01

+++ 0x01-variables_if_else_while/1-last_digit.c
@@ -17,18 +17,18 @@ int main(void)
 	srand(time(0));
 	n = rand() - RAND_MAX / 2;
 
-	int last_digit = n % 10;
+	float last_digit = n % 10;
 
 	if (last_digit > 5)
 	{
-		printf("Last digit of %d is %d and is greater than 5\n", n,
+		printf("Last digit of %d is %f and is greater than 5\n", n,
 				last_digit);
 	} else if (last_digit == 0)
 	{
-		printf("Last digit of %d is %d and is 0\n", n, last_digit);
+		printf("Last digit of %d is %f and is 0\n", n, last_digit);
 	} else if (last_digit < 6 && last_digit != 0)
 	{
-		printf("Last digit of %d is %d and is less than 6 and not 0\n",
+		printf("Last digit of %d is %f and is less than 6 and not 0\n",
 				n, last_digit);
 	}
